% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{asBanksyObject}
\alias{asBanksyObject}
\title{Convert a SingleCellExperiment to BanksyObject}
\usage{
asBanksyObject(
  se,
  expr.assay = "counts",
  features = NULL,
  coord.colnames = c("sdimx", "sdimy")
)
}
\arguments{
\item{se}{object of class SingleCellExperiment or SpatialExperiment}

\item{expr.assay}{expression assay to extract}

\item{features}{features to extract from expression assay}

\item{coord.colnames}{colnames of expression in metadata from SCE}
}
\value{
BanksyObject
}
\description{
Convert a SingleCellExperiment to BanksyObject
}
\examples{
data <- simulateDataset()
sce <- SingleCellExperiment::SingleCellExperiment(
    assays = list(counts = data$gcm), colData = data$locs)
bank <- asBanksyObject(sce, expr.assay = 'counts', 
                            coord.colnames = c('x', 'y'))
}
