% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BanksyObject.R
\docType{class}
\name{BanksyObject-class}
\alias{BanksyObject-class}
\alias{BanksyObject}
\alias{own.expr}
\alias{own.expr,BanksyObject-method}
\alias{nbr.expr}
\alias{nbr.expr,BanksyObject-method}
\alias{harmonics}
\alias{harmonics,BanksyObject-method}
\alias{custom.expr}
\alias{custom.expr,BanksyObject-method}
\alias{cell.locs}
\alias{cell.locs,BanksyObject-method}
\alias{meta.data}
\alias{meta.data,BanksyObject-method}
\alias{reduction}
\alias{reduction,BanksyObject-method}
\alias{clust.names}
\alias{clust.names,BanksyObject-method}
\alias{reduction.names}
\alias{reduction.names,BanksyObject-method}
\alias{own.expr<-}
\alias{own.expr<-,BanksyObject,assay-method}
\alias{nbr.expr<-}
\alias{nbr.expr<-,BanksyObject,assay-method}
\alias{harmonics<-}
\alias{harmonics<-,BanksyObject,assay-method}
\alias{custom.expr<-}
\alias{custom.expr<-,BanksyObject,assay-method}
\alias{cell.locs<-}
\alias{cell.locs<-,BanksyObject,assay-method}
\alias{meta.data<-}
\alias{meta.data<-,BanksyObject,data.frame-method}
\alias{reduction<-}
\alias{reduction<-,BanksyObject,list-method}
\alias{show,BanksyObject-method}
\alias{head,BanksyObject-method}
\title{S4 class BanksyObject}
\usage{
BanksyObject(
  own.expr = NULL,
  nbr.expr = NULL,
  harmonics = NULL,
  custom.expr = NULL,
  cell.locs = NULL,
  meta.data = NULL,
  reduction = NULL,
  genes.filter = "intersect",
  min.cells.expressed = -1
)

own.expr(object)

\S4method{own.expr}{BanksyObject}(object)

nbr.expr(object)

\S4method{nbr.expr}{BanksyObject}(object)

harmonics(object)

\S4method{harmonics}{BanksyObject}(object)

custom.expr(object)

\S4method{custom.expr}{BanksyObject}(object)

cell.locs(object)

\S4method{cell.locs}{BanksyObject}(object)

meta.data(object)

\S4method{meta.data}{BanksyObject}(object)

reduction(object)

\S4method{reduction}{BanksyObject}(object)

clust.names(object)

\S4method{clust.names}{BanksyObject}(object)

reduction.names(object)

\S4method{reduction.names}{BanksyObject}(object)

own.expr(object) <- value

\S4method{own.expr}{BanksyObject,assay}(object) <- value

nbr.expr(object) <- value

\S4method{nbr.expr}{BanksyObject,assay}(object) <- value

harmonics(object) <- value

\S4method{harmonics}{BanksyObject,assay}(object) <- value

custom.expr(object) <- value

\S4method{custom.expr}{BanksyObject,assay}(object) <- value

cell.locs(object) <- value

\S4method{cell.locs}{BanksyObject,assay}(object) <- value

meta.data(object) <- value

\S4method{meta.data}{BanksyObject,data.frame}(object) <- value

reduction(object) <- value

\S4method{reduction}{BanksyObject,list}(object) <- value

\S4method{show}{BanksyObject}(object)

\S4method{head}{BanksyObject}(x, n = 5)
}
\arguments{
\item{own.expr}{own expression}

\item{nbr.expr}{neighbour expression}

\item{harmonics}{azimuthal fourier harmonics}

\item{custom.expr}{custom expression}

\item{cell.locs}{cell locations}

\item{meta.data}{metadata}

\item{reduction}{dimension reductions}

\item{genes.filter}{intersect or union}

\item{min.cells.expressed}{min number of cells a gene is expressed in
(to be used with intersect)}

\item{object}{BanksyObject}

\item{value}{to be assigned}

\item{x}{BanksyObject}

\item{n}{Number of records to display}
}
\value{
BanksyObject
}
\description{
S4 class BanksyObject
}
\section{Functions}{
\itemize{
\item \code{own.expr()}: getter own.expr

\item \code{own.expr(BanksyObject)}: getter own.expr

\item \code{nbr.expr()}: getter nbr.expr

\item \code{nbr.expr(BanksyObject)}: getter nbr.expr

\item \code{harmonics()}: getter harmonics

\item \code{harmonics(BanksyObject)}: getter harmonics

\item \code{custom.expr()}: getter custom.expr

\item \code{custom.expr(BanksyObject)}: getter custom.expr

\item \code{cell.locs()}: getter cell.locs

\item \code{cell.locs(BanksyObject)}: getter cell.locs

\item \code{meta.data()}: getter meta.data

\item \code{meta.data(BanksyObject)}: getter meta.data

\item \code{reduction()}: getter reduction

\item \code{reduction(BanksyObject)}: getter reduction

\item \code{clust.names()}: getter clust.names

\item \code{clust.names(BanksyObject)}: getter clust.names

\item \code{reduction.names()}: getter reduction.names

\item \code{reduction.names(BanksyObject)}: getter reduction.names

\item \code{own.expr(object) <- value}: setter for own.expr

\item \code{own.expr(object = BanksyObject) <- value}: setter for own.expr

\item \code{nbr.expr(object) <- value}: setter for nbr.expr

\item \code{nbr.expr(object = BanksyObject) <- value}: setter for nbr.expr

\item \code{harmonics(object) <- value}: setter for harmonics

\item \code{harmonics(object = BanksyObject) <- value}: setter for harmonics

\item \code{custom.expr(object) <- value}: setter for custom.expr

\item \code{custom.expr(object = BanksyObject) <- value}: setter for custom.expr

\item \code{cell.locs(object) <- value}: setter for cell.locs

\item \code{cell.locs(object = BanksyObject) <- value}: setter for cell.locs

\item \code{meta.data(object) <- value}: setter for meta.data

\item \code{meta.data(object = BanksyObject) <- value}: setter for meta.data

\item \code{reduction(object) <- value}: setter for reduction

\item \code{reduction(object = BanksyObject) <- value}: setter for reduction

\item \code{show(BanksyObject)}: show BanksyObject

\item \code{head(BanksyObject)}: header for BanksyObject

}}
\section{Slots}{

\describe{
\item{\code{own.expr}}{own expression}

\item{\code{nbr.expr}}{neighbour expression}

\item{\code{harmonics}}{azimuthal fourier harmonics}

\item{\code{custom.expr}}{custom expression}

\item{\code{cell.locs}}{cell locations}

\item{\code{meta.data}}{metadata}

\item{\code{reduction}}{dimension reductions}
}}

\examples{
d <- simulateDataset()
bank <- BanksyObject(
    own.expr = d$gcm, 
    cell.locs = d$locs, 
    meta.data = d$meta
    )
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
mat <- own.expr(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
mat <- nbr.expr(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
lst <- harmonics(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
mat <- custom.expr(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
df <- cell.locs(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
df <- meta.data(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
lst <- reduction(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
vec <- clust.names(bank)
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
vec <- reduction.names(bank)
}
