% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cerebellum}
\alias{cerebellum}
\title{Mouse cerebellum Slide-seq V2 data}
\format{
A list with 2 entries:
\describe{
\item{expression}{(matrix) gene expression matrix}
\item{locations}{(data.frame) cell centroids in 2D}
}
}
\source{
\url{https://singlecell.broadinstitute.org/single_cell/study/SCP948}
}
\usage{
data(cerebellum)
}
\description{
This dataset comprise Slide-seq V2 profiling of cells of the mouse
cerebellum. Gene expression and cell centroids for 9,953 cells and 8,365
genes in 2 spatial dimensions are provided.
}
\keyword{datasets}
