% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{dim.BanksyLazy}
\alias{dim.BanksyLazy}
\title{S3 dim method for BanksyLazy operator}
\usage{
\method{dim}{BanksyLazy}(x)
}
\arguments{
\item{x}{A BanksyLazy object}
}
\value{
Integer vector of length 2: c(2*n_genes, n_cells)
}
\description{
S3 dim method for BanksyLazy operator
}
\keyword{internal}
