% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlpfc151673}
\alias{dlpfc151673}
\title{Human dorsolateral prefrontal cortex 10x Visium data (sample 151673)}
\format{
A list with 3 entries:
\describe{
\item{expression}{(matrix) gene expression matrix}
\item{locations}{(data.frame) spot locations in 2D}
}
}
\source{
\url{http://research.libd.org/spatialLIBD/}
}
\usage{
dlpfc151673
}
\description{
This dataset comprises 10x Visium profiling of the human DLPFC (Maynard et
al. (2021) Transcriptome-scale spatial gene expression in the human
dorsolateral prefrontal cortex. doi.org: 10.1038/s41593-020-00787-0) for
subject 3 sample 151673. Gene expression and spot locations for 3,639 spots
and top 5000 highly variable genes in 2 spatial dimensions are provided.
}
\keyword{datasets}
