% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getARI}
\alias{getARI}
\title{Calculate adjusted rand index for harmonized clusters}
\usage{
getARI(bank, digits = 3)
}
\arguments{
\item{bank}{BanksyObject with harmonized clusters}

\item{digits}{number of digits to round ARI to}
}
\value{
matrix of ARI
}
\description{
Calculate adjusted rand index for harmonized clusters
}
