% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{getMarkerGenes}
\alias{getMarkerGenes}
\title{Get marker genes}
\usage{
getMarkerGenes(
  gobject = NULL,
  sweepresults = NULL,
  pairConnectedOutput = NULL,
  numMarkers = 2,
  upper_clip = 2,
  lower_clip = -2,
  markerMethod = c("scran", "scran_and_gini", "gini")
)
}
\arguments{
\item{gobject}{Giotto object}

\item{sweepresults}{returned by SweepBanksy}

\item{pairConnectedOutput}{returned by gridConnect}

\item{numMarkers}{number of markers}

\item{upper_clip}{upp}

\item{lower_clip}{low}

\item{markerMethod}{method used}
}
\value{
List of marker genes
}
\description{
Get marker genes
}
