% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{gridConnect}
\alias{gridConnect}
\title{Connects the colours in the grid of the parameter sweep}
\usage{
gridConnect(sweepresults, rootnode = c(5, 2))
}
\arguments{
\item{sweepresults}{returned by SweepResults}

\item{rootnode}{root node used}
}
\value{
PairedConnectedOutput
}
\description{
Connects the colours in the grid of the parameter sweep
}
