% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypothalamus}
\alias{hypothalamus}
\title{Mouse Hypothalamus MERFISH data}
\format{
A list with 3 entries:
\describe{
\item{expression}{(matrix) gene expression matrix}
\item{locations}{(data.frame) cell centroids in 3D}
\item{metadata}{(data.frame) metadata for cells, including cell class,
  neuronal cluster, animal ID, behavioral class}
}
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.8t8s248}
}
\usage{
hypothalamus
}
\description{
This dataset comprises MERFISH profiling of cells in the pre-optic region
of the mouse hypothalamus (Moffitt et al. (2018), Molecular, spatial, and
functional single-cell profiling of the hypothalamic preoptic region.
doi: 10.1126/science.aau5324). The data is subset to 2 z-planes from a single
animal (Animal 1, bregmas 0.21 and 0.26). Gene expression, cell centroids
and metadata for 11,162 cells and 161 genes in 3 spatial dimensions are
provided.
}
\keyword{datasets}
