% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{organoid}
\alias{organoid}
\title{Mouse brain organoid MERFISH data}
\format{
A list with 2 entries:
\describe{
\item{expression}{(list) a list with 3 entries, each a gene expression
  matrix}
\item{locations}{(list) a list with 3 entries with corresponding cell
  centroids in 2D)}
}
}
\usage{
organoid
}
\description{
This data consists of 3 separate datasets comprising MERFISH profiling of
mouse brain organoids. Gene expression and cell centroids for 3,600 cells and
at least 557 genes in 2 spatial dimensions are provided.
}
\keyword{datasets}
