% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotAlluvia}
\alias{plotAlluvia}
\title{Plot Alluvial plot for different cluster assignments}
\usage{
plotAlluvia(bank)
}
\arguments{
\item{bank}{BanksyObject}
}
\value{
Alluvial plot
}
\description{
Plot Alluvial plot for different cluster assignments
}
