% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotARI}
\alias{plotARI}
\title{Plot ARI matrix}
\usage{
plotARI(bank, col.low = "white", col.high = "red", label = TRUE, digits = 3)
}
\arguments{
\item{bank}{BanksyObject with harmonised clusterings}

\item{col.low}{color gradient low}

\item{col.high}{color gradient high}

\item{label}{TRUE if to label plot with ARI}

\item{digits}{number of digits to round ARI to}
}
\description{
Plot ARI matrix
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunPCA(bank, lambda = 0.2)
bank <- ClusterBanksy(bank, lambda = 0.2, npcs = 20, k.neighbors = 50, resolution = c(0.5,1.5))
plotARI(bank)

}
