% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHeatmap_banksy}
\alias{plotHeatmap_banksy}
\title{Plot Banksy Heatmap}
\usage{
plotHeatmap_banksy(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  plotValuesRange = NULL,
  genes,
  cluster_column = NULL,
  cluster_order = c("size", "correlation", "custom"),
  cluster_custom_order = NULL,
  cluster_color_code = NULL,
  cluster_cor_method = "pearson",
  cluster_hclust_method = "ward.D",
  gene_order = c("correlation", "custom"),
  gene_custom_order = NULL,
  gene_cor_method = "pearson",
  gene_hclust_method = "complete",
  show_values = c("rescaled", "z-scaled", "original"),
  size_vertical_lines = 1.1,
  gradient_colors = c("blue", "yellow", "red"),
  gene_label_selection = NULL,
  axis_text_y_size = NULL,
  legend_nrows = 1,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotHeatmap"
)
}
\arguments{
\item{gobject}{Giotto object}

\item{expression_values}{one of normalized, scaled, custom}

\item{plotValuesRange}{plot value range}

\item{genes}{genes to use}

\item{cluster_column}{column to cluster by}

\item{cluster_order}{custom order for clusters}

\item{cluster_custom_order}{custom cluster order}

\item{cluster_color_code}{custom colors for clusters}

\item{cluster_cor_method}{method for cluster correlation}

\item{cluster_hclust_method}{method for hierarchical clustering of clusters}

\item{gene_order}{method to determine gene order}

\item{gene_custom_order}{custom order for genes}

\item{gene_cor_method}{method for gene correlation}

\item{gene_hclust_method}{method for hierarchical clustering of genes}

\item{show_values}{show values or not}

\item{size_vertical_lines}{size of vertical lines}

\item{gradient_colors}{grad colours}

\item{gene_label_selection}{select genes}

\item{axis_text_y_size}{size of y axis text}

\item{legend_nrows}{nrows}

\item{show_plot}{show plot or not}

\item{return_plot}{return plot or not}

\item{save_plot}{save plot or not}

\item{save_param}{save parameters}

\item{default_save_name}{save name}
}
\value{
Combined heatmap plot
}
\description{
Plot Banksy Heatmap
}
