% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMetaDataHeatmap_vs}
\alias{plotMetaDataHeatmap_vs}
\title{Plots expression heatmaps}
\usage{
plotMetaDataHeatmap_vs(
  gobject,
  expression_values = c("normalized", "scaled", "custom"),
  metadata_cols = NULL,
  selected_genes = NULL,
  first_meta_col = NULL,
  second_meta_col = NULL,
  show_values = c("zscores", "rescaled", "original", "zscores_rescaled"),
  custom_cluster_order = NULL,
  clus_cor_method = "pearson",
  clus_cluster_method = "complete",
  custom_gene_order = NULL,
  gene_cor_method = "pearson",
  gene_cluster_method = "complete",
  gradient_color = c("blue", "white", "red"),
  gradient_midpoint = 0.5,
  gradient_limits = NULL,
  x_text_size = 10,
  x_text_angle = 45,
  y_text_size = 10,
  strip_text_size = 8,
  show_plot = NA,
  return_plot = NA,
  save_plot = NA,
  save_param = list(),
  default_save_name = "plotMetaDataHeatmap"
)
}
\arguments{
\item{gobject}{Giotto object}

\item{expression_values}{one of normalized, scaled, custom}

\item{metadata_cols}{mdata}

\item{selected_genes}{genes to use}

\item{first_meta_col}{1st meta col}

\item{second_meta_col}{2nd meta col}

\item{show_values}{one of zscores, rescaled, original, zscores_rescaled}

\item{custom_cluster_order}{cell clustering order}

\item{clus_cor_method}{cell correlation method}

\item{clus_cluster_method}{cell clustering method}

\item{custom_gene_order}{gene order}

\item{gene_cor_method}{gene correlation method}

\item{gene_cluster_method}{gene clustering method}

\item{gradient_color}{grad colour}

\item{gradient_midpoint}{grad midpoint}

\item{gradient_limits}{gradlimits}

\item{x_text_size}{x text size}

\item{x_text_angle}{x text angle}

\item{y_text_size}{y text size}

\item{strip_text_size}{strip text size}

\item{show_plot}{show plot}

\item{return_plot}{return plot}

\item{save_plot}{save plot}

\item{save_param}{save parameters}

\item{default_save_name}{default saving name}
}
\value{
Heatmap
}
\description{
Plots expression heatmaps
}
