% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{plotScree}
\alias{plotScree}
\title{Plots scree plot for PCA}
\usage{
plotScree(bank, lambda)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{numeric}
}
\value{
Scree plot
}
\description{
Plots scree plot for PCA
}
