% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpatDimPlots}
\alias{plotSpatDimPlots}
\title{Plots spatial dimension plots}
\usage{
plotSpatDimPlots(
  gobject,
  sweepresults,
  pairConnectedOutput,
  upper_clip = 3,
  lower_clip = -3,
  upper_clip2 = 2,
  lower_clip2 = -2,
  markerList = NULL,
  point_size = 0.5,
  full_color_palate = NULL,
  custom_rowwise_ind = NULL,
  additional_string = "",
  show_plot = TRUE,
  plot_spatial_dim_plots = TRUE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{sweepresults}{sweep results}

\item{pairConnectedOutput}{paired connected output}

\item{upper_clip}{upper clip}

\item{lower_clip}{lower clip}

\item{upper_clip2}{upper clip}

\item{lower_clip2}{lower clip}

\item{markerList}{marker list}

\item{point_size}{point}

\item{full_color_palate}{color palete}

\item{custom_rowwise_ind}{custom row wise indices}

\item{additional_string}{additional string for name}

\item{plot_spatial_dim_plots}{whether to plot or not}
}
\value{
Spatial dim plot
}
\description{
Plots spatial dimension plots
}
