% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSpatial}
\alias{plotSpatial}
\title{Plot Spatial dims}
\usage{
plotSpatial(
  bank,
  dataset = NULL,
  by = NA,
  type = c("discrete", "continuous"),
  pt.size = 0.5,
  pt.alpha = 0.7,
  col.midpoint = NULL,
  col.lowpoint = NULL,
  col.highpoint = NULL,
  col.low = "blue",
  col.mid = "gray95",
  col.high = "red",
  na.value = "gray",
  col.discrete = NULL,
  main = NULL,
  main.size = 5,
  legend = TRUE,
  legend.text.size = 6,
  legend.pt.size = 3
)
}
\arguments{
\item{bank}{BanksyObject}

\item{dataset}{if multiple dataset are run}

\item{by}{groupings for cells}

\item{type}{type of groupings (one of 'discrete' or 'continuous')}

\item{pt.size}{size of points}

\item{pt.alpha}{transparency of points}

\item{col.midpoint}{for continuous labels - midpoint of color gradient}

\item{col.lowpoint}{for continuous labels - low limit of scale}

\item{col.highpoint}{for continuous labels - high limit of scale}

\item{col.low}{for continuous labels - color gradient low}

\item{col.mid}{for continuous labels - color gradient mid}

\item{col.high}{for continuous labels - color gradient high}

\item{na.value}{for continuous labels - colors to use for missing values}

\item{col.discrete}{for discrete labels - colors for groupings}

\item{main}{title}

\item{main.size}{size of title}

\item{legend}{show legend}

\item{legend.text.size}{size of legend text}

\item{legend.pt.size}{size of legend point}
}
\value{
plot
}
\description{
Plot Spatial dims
}
