% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpatialDims}
\alias{plotSpatialDims}
\title{Plot Spatial dims}
\usage{
plotSpatialDims(bank, params, col = NULL, main = NULL, size = 1)
}
\arguments{
\item{bank}{BanksyObject}

\item{params}{clustering run to plot by}

\item{col}{colours}

\item{main}{title}

\item{size}{size of points}
}
\description{
Plot Spatial dims
}
