% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSpatialFeatures}
\alias{plotSpatialFeatures}
\title{Wrapper for plotSpatial}
\usage{
plotSpatialFeatures(bank, dataset = NULL, by, type, nrow, ncol, ...)
}
\arguments{
\item{bank}{BanksyObject}

\item{dataset}{if multiple datasets are used}

\item{by}{features to plot}

\item{type}{type corresponding to features}

\item{nrow}{nrow for grob}

\item{ncol}{ncol for grob}

\item{...}{parameters to pass to plotSpatial}
}
\description{
Wrapper for plotSpatial
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunPCA(bank, lambda = 0.2)
bank <- ClusterBanksy(bank, lambda = 0.2, npcs = 20, k.neighbors = 50, resolution = 0.8)
plotSpatialFeatures(bank, by = c('Label', clust.names(bank)), 
  type = rep('discrete', 2), nrow = 1, ncol = 2, pt.size = 2)

}
