% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP}
\usage{
plotUMAP(
  bank,
  params,
  col = NULL,
  legend = TRUE,
  main = NULL,
  pt.size = 1,
  main.size = 1
)
}
\arguments{
\item{bank}{BanksyObject}

\item{params}{clustering run to plot by}

\item{col}{cluster colours}

\item{legend}{cluster legend}

\item{main}{title}

\item{pt.size}{size of points}

\item{main.size}{size of title}
}
\description{
Plot UMAP
}
