% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{createGiottoInstructions}
\alias{createGiottoObject}
\alias{normalizeGiotto}
\alias{copy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Giotto}{\code{\link[Giotto]{createGiottoInstructions}}, \code{\link[Giotto]{createGiottoObject}}, \code{\link[Giotto]{normalizeGiotto}}}

  \item{data.table}{\code{\link[data.table]{copy}}}
}}

