% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{RunAllWilcox}
\alias{RunAllWilcox}
\title{Wilcoxon two-sided tests to find markers for all identity class (based on Seurat)}
\usage{
RunAllWilcox(
  bank,
  assay = "own.expr",
  by = NULL,
  q.val.thres = 0.01,
  express.thres = 1,
  only.pos = TRUE,
  pseudocount.use = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{bank}{BanksyObject with unscaled expression}

\item{assay}{assay to find markers with
\itemize{
  \item{own.expr: own expression (default)}
  \item{nbr.expr: neighbor expression}
}}

\item{by}{metadata column to define identity classes (default uses latest clustering)}

\item{q.val.thres}{q-value threshold to filter out markers (default: 0.01)}

\item{express.thres}{count threshold for gene to be considered expressed in cell (default: 1)}

\item{only.pos}{return markers with positive log2 fold change (i.e. for identity class 1)}

\item{pseudocount.use}{pseudocount for fold change computation}

\item{verbose}{verbose}
}
\value{
dataframe of markers
}
\description{
Wilcoxon two-sided tests to find markers for all identity class (based on Seurat)
}
\details{
returns seven columns:
\itemize{
  \item{p.val: pvalue returned by the Wilcoxon rank-sum test}
  \item{pct.1: percentage of cells in id1 expressing gene (based on express.thres)}
  \item{pct.2: percentage of cells in id2 expressing gene (based on express.thres)}
  \item{log.FC: log2 fold change between mean of id1 and id2 with pseudocount}
  \item{q.val: qvalue returned by correction of pvalues}
  \item{gene: marker name}
  \item{cluster: cluster for marker}
}
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
markers <- RunWilcox(bank, by = 'Label')

}
