% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{RunBanksy}
\alias{RunBanksy}
\title{This function takes the gene-cell matrix and cell locations, and runs banksy.}
\usage{
RunBanksy(
  gcm,
  locs,
  sigma = 3,
  lambda = 0.25,
  alpha = 0.05,
  pcs = 15,
  k_expr = 30,
  res = 0.5,
  kspatial = 1000,
  normalizeColumns = TRUE,
  normalizeColumnsTo = 100,
  minimumTotalCounts = 10,
  lieden_iters = -1,
  zScaleRows = TRUE,
  zScaleBeforeAveraging = FALSE,
  zScaleOwnAfterAveraging = TRUE,
  zScaleNbrAfterAveraging = TRUE,
  zScaleDatasetsIndividually = TRUE,
  multiple_datasets = FALSE,
  cells_split_by_dataset = NULL,
  dimensions = "all",
  clustering_name = "banksy",
  k_geom = 10,
  banksyMatrices = NULL,
  spatialMode = c("kNN_r", "kNearestNeighbours", "kNN_unif", "kNN_rank", "rNN_gauss"),
  compute_own_and_nbr_umap = FALSE,
  compute_3D_umap = FALSE,
  instructions,
  ...
)
}
\arguments{
\item{gcm}{gene-cell matrix, rows are features, columns are objects}

\item{locs}{cell locations with object locations: x and y (and possibly z)
coordinates. x and y are necessary, and must have at least two different
locations. If data is in a single plane, it must be in a single z plane.
The value of this z plane variable may be provided, but it is not necessary.
There needs to be a column labeled cell_ID, which contains strings that match
the column names of gcm, in the same order as the columns of gcm. locs should
be a data.table with rows and columns as follows:
sdimx     sdimy  sdimz    cell_ID}

\item{sigma}{sigma}

\item{lambda}{lambda}

\item{alpha}{alpha}

\item{pcs}{pcs}

\item{k_expr}{k_expr for shared neighbourhood clustering}

\item{res}{res}

\item{kspatial}{kspatial}

\item{normalizeColumns}{norm}

\item{normalizeColumnsTo}{norm}

\item{minimumTotalCounts}{min}

\item{lieden_iters}{lie}

\item{zScaleRows}{z}

\item{zScaleBeforeAveraging}{z}

\item{zScaleOwnAfterAveraging}{z}

\item{zScaleNbrAfterAveraging}{z}

\item{zScaleDatasetsIndividually}{z}

\item{multiple_datasets}{multiple}

\item{cells_split_by_dataset}{no}

\item{dimensions}{dims}

\item{clustering_name}{name}

\item{k_geom}{k}

\item{banksyMatrices}{mat}

\item{spatialMode}{mode}

\item{compute_own_and_nbr_umap}{no}

\item{compute_3D_umap}{no}

\item{instructions}{instructions}

\item{...}{more params}
}
\value{
List of clustering results
}
\description{
This function takes the gene-cell matrix and cell locations, and runs banksy.
}
