% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{RunPCA}
\alias{RunPCA}
\title{Run PCA}
\usage{
RunPCA(bank, lambda, npcs = 30)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{(numeric vector) spatial weighting parameter}

\item{npcs}{(numeric) number of principal components to compute}
}
\value{
BanksyObject with PCA
}
\description{
Run PCA
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunPCA(bank, lambda = 0.2)

}
