% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{SweepBanksy}
\alias{SweepBanksy}
\title{Sweep gene-cell matrix}
\usage{
SweepBanksy(
  gcm,
  locs,
  rowparam = "lam",
  rowvalues = c(0, 0.25, 0.5),
  colparam = "res",
  colvalues = c(0.5, 1),
  sigma = 3,
  lambda = 0.4,
  alpha = 0.05,
  pcs = 15,
  k_expr = 30,
  res = 0.5,
  kspatial = 1000,
  k_geom = 10,
  normalizeColumns = TRUE,
  normalizeColumnsTo = 100,
  minimumTotalCounts = 0,
  lieden_iters = -1,
  zScaleRows = TRUE,
  zScaleBeforeAveraging = FALSE,
  zScaleOwnAfterAveraging = TRUE,
  zScaleNbrAfterAveraging = TRUE,
  zScaleDatasetsIndividually = TRUE,
  multiple_datasets = FALSE,
  cells_split_by_dataset = NULL,
  dimensions = "all",
  spatialMode = c("kNN_r", "kNearestNeighbours", "kNN_unif", "kNN_rank", "rNN_gauss"),
  banksyMatrices = NULL,
  compute_own_and_nbr_umap = FALSE,
  compute_3D_umap = FALSE,
  instructions
)
}
\arguments{
\item{gcm}{gene-cell matrix, rows are features, columns are objects}

\item{locs}{cell locations with object locations: x and y (and possibly z)
coordinates. x and y are necessary, and must have at least two different
locations. If data is in a single plane, it must be in a single z plane.
The value of this z plane variable may be provided, but it is not necessary.
There needs to be a column labeled cell_ID, which contains strings that match
the column names of gcm, in the same order as the columns of gcm. locs should
be a data.table with rows and columns as follows:
sdimx     sdimy  sdimz    cell_ID}

\item{rowparam}{rowparam}

\item{rowvalues}{rowvals}

\item{colparam}{colparam}

\item{colvalues}{colvals}

\item{sigma}{sigma}

\item{lambda}{lambda}

\item{alpha}{alpha}

\item{pcs}{pcs}

\item{k_expr}{k_expr}

\item{res}{res}

\item{kspatial}{kspatial}

\item{k_geom}{k_geom}

\item{normalizeColumns}{norm}

\item{normalizeColumnsTo}{norm}

\item{minimumTotalCounts}{min}

\item{lieden_iters}{niters}

\item{zScaleRows}{z}

\item{zScaleBeforeAveraging}{z}

\item{zScaleOwnAfterAveraging}{z}

\item{zScaleNbrAfterAveraging}{z}

\item{zScaleDatasetsIndividually}{z}

\item{multiple_datasets}{m}

\item{cells_split_by_dataset}{cells}

\item{dimensions}{dims}

\item{spatialMode}{spatial}

\item{banksyMatrices}{bank}

\item{compute_own_and_nbr_umap}{no}

\item{compute_3D_umap}{no}

\item{instructions}{instrs}
}
\value{
Sweep List
}
\description{
Sweep gene-cell matrix
}
