% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{updateGobjectUmapClust}
\alias{updateGobjectUmapClust}
\title{Updates Giotto object UMAP clusters with sweep results and connected output}
\usage{
updateGobjectUmapClust(
  gobject,
  sweepresults,
  pairConnectedOutput,
  compute_own_and_nbr_umap = FALSE,
  compute_3D_umap = FALSE
)
}
\arguments{
\item{gobject}{Giotto object}

\item{sweepresults}{returned by SweepResults}

\item{pairConnectedOutput}{returned by gridConnect}

\item{compute_own_and_nbr_umap}{no}

\item{compute_3D_umap}{no}
}
\value{
updated Giotto object
}
\description{
Updates Giotto object UMAP clusters with sweep results and connected output
}
