% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visualcortex}
\alias{visualcortex}
\title{Mouse visual cortex SeqFISH data}
\format{
A list with 3 entries:
\describe{
\item{expression}{(matrix) gene expression matrix}
\item{locations}{(data.frame) cell centroids in 2D}
\item{metadata}{(data.frame) metadata for cells, including HMRF domain}
\item{spatialgenes}{(character) spatial genes for HMRF analysis}
}
}
\source{
\url{https://bitbucket.org/qzhudfci/smfishhmrf-r/src}
}
\usage{
data(visualcortex)
}
\description{
This dataset comprises SeqFISH profiling of cells in the mouse visual cortex
(Qian Zhu et al. (2018) Identification of spatially associated subpopulations
by combining scRNA-seq and sequential florescence in situ hybridization data.
doi: 10.1038/nbt.4260). Gene expression, cell centroids and metadata are
provided for 1,597 cells, 125 genes in 2 spatial dimensions are provided.
In addition, the spatial genes used for HMRF analysis in the
reference are included.
}
\keyword{datasets}
