% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BanksyObject.R
\docType{class}
\name{BanksyObject-class}
\alias{BanksyObject-class}
\alias{BanksyObject}
\alias{own.expr}
\alias{own.expr,BanksyObject-method}
\alias{nbr.expr}
\alias{nbr.expr,BanksyObject-method}
\alias{own.norm.scaled.expr}
\alias{own.norm.scaled.expr,BanksyObject-method}
\alias{nbr.norm.scaled.expr}
\alias{nbr.norm.scaled.expr,BanksyObject-method}
\alias{custom.expr}
\alias{custom.expr,BanksyObject-method}
\alias{cell.locs}
\alias{cell.locs,BanksyObject-method}
\alias{meta.data}
\alias{meta.data,BanksyObject-method}
\alias{dim.reduction}
\alias{dim.reduction,BanksyObject-method}
\alias{own.expr<-}
\alias{own.expr<-,BanksyObject,assay-method}
\alias{nbr.expr<-}
\alias{nbr.expr<-,BanksyObject,assay-method}
\alias{own.norm.scaled.expr<-}
\alias{own.norm.scaled.expr<-,BanksyObject,assay-method}
\alias{nbr.norm.scaled.expr<-}
\alias{nbr.norm.scaled.expr<-,BanksyObject,assay-method}
\alias{custom.expr<-}
\alias{custom.expr<-,BanksyObject,assay-method}
\alias{cell.locs<-}
\alias{cell.locs<-,BanksyObject,assay-method}
\alias{meta.data<-}
\alias{meta.data<-,BanksyObject,data.frame-method}
\alias{dim.reduction<-}
\alias{dim.reduction<-,BanksyObject,list-method}
\alias{head}
\alias{head,BanksyObject-class}
\title{S4 class BanksyObject}
\usage{
BanksyObject(
  own.expr = NULL,
  nbr.expr = NULL,
  own.norm.scaled.expr = NULL,
  nbr.norm.scaled.expr = NULL,
  custom.expr = NULL,
  cell.locs = NULL,
  meta.data = data.frame(),
  dim.reduction = NULL
)

own.expr(object)

\S4method{own.expr}{BanksyObject}(object)

nbr.expr(object)

\S4method{nbr.expr}{BanksyObject}(object)

own.norm.scaled.expr(object)

\S4method{own.norm.scaled.expr}{BanksyObject}(object)

nbr.norm.scaled.expr(object)

\S4method{nbr.norm.scaled.expr}{BanksyObject}(object)

custom.expr(object)

\S4method{custom.expr}{BanksyObject}(object)

cell.locs(object)

\S4method{cell.locs}{BanksyObject}(object)

meta.data(object)

\S4method{meta.data}{BanksyObject}(object)

dim.reduction(object)

\S4method{dim.reduction}{BanksyObject}(object)

own.expr(object) <- value

\S4method{own.expr}{BanksyObject,assay}(object) <- value

nbr.expr(object) <- value

\S4method{nbr.expr}{BanksyObject,assay}(object) <- value

own.norm.scaled.expr(object) <- value

\S4method{own.norm.scaled.expr}{BanksyObject,assay}(object) <- value

nbr.norm.scaled.expr(object) <- value

\S4method{nbr.norm.scaled.expr}{BanksyObject,assay}(object) <- value

custom.expr(object) <- value

\S4method{custom.expr}{BanksyObject,assay}(object) <- value

cell.locs(object) <- value

\S4method{cell.locs}{BanksyObject,assay}(object) <- value

meta.data(object) <- value

\S4method{meta.data}{BanksyObject,data.frame}(object) <- value

dim.reduction(object) <- value

\S4method{dim.reduction}{BanksyObject,list}(object) <- value

head(x, ...)
}
\arguments{
\item{own.expr}{own expression}

\item{nbr.expr}{neighbour expression}

\item{own.norm.scaled.expr}{own norm/scaled expression}

\item{nbr.norm.scaled.expr}{neighbour norm/scaled expression}

\item{custom.expr}{custom expression}

\item{cell.locs}{cell locations}

\item{meta.data}{metadata}

\item{dim.reduction}{dimension reductions}

\item{object}{BanksyObject}

\item{value}{to be assigned}

\item{x}{BanksyObject}
}
\value{
BanksyObject
}
\description{
S4 class BanksyObject
}
\section{Functions}{
\itemize{
\item \code{own.expr}: getter own.expr

\item \code{own.expr,BanksyObject-method}: getter own.expr

\item \code{nbr.expr}: getter nbr.expr

\item \code{nbr.expr,BanksyObject-method}: getter nbr.expr

\item \code{own.norm.scaled.expr}: getter own.norm.scaled.expr

\item \code{own.norm.scaled.expr,BanksyObject-method}: getter own.norm.scaled.expr

\item \code{nbr.norm.scaled.expr}: getter nbr.norm.scaled.expr

\item \code{nbr.norm.scaled.expr,BanksyObject-method}: getter nbr.norm.scaled.expr

\item \code{custom.expr}: getter custom.expr

\item \code{custom.expr,BanksyObject-method}: getter custom.expr

\item \code{cell.locs}: getter cell.locs

\item \code{cell.locs,BanksyObject-method}: getter cell.locs

\item \code{meta.data}: getter meta.data

\item \code{meta.data,BanksyObject-method}: getter meta.data

\item \code{dim.reduction}: getter dim.reduction

\item \code{dim.reduction,BanksyObject-method}: getter dim.reduction

\item \code{own.expr<-}: setter for own.expr

\item \code{own.expr<-,BanksyObject,assay-method}: setter for own.expr

\item \code{nbr.expr<-}: setter for nbr.expr

\item \code{nbr.expr<-,BanksyObject,assay-method}: setter for nbr.expr

\item \code{own.norm.scaled.expr<-}: setter for own.norm.scaled.expr

\item \code{own.norm.scaled.expr<-,BanksyObject,assay-method}: setter for own.norm.scaled.expr

\item \code{nbr.norm.scaled.expr<-}: setter for nbr.norm.scaled.expr

\item \code{nbr.norm.scaled.expr<-,BanksyObject,assay-method}: setter for nbr.norm.scaled.expr

\item \code{custom.expr<-}: setter for custom.expr

\item \code{custom.expr<-,BanksyObject,assay-method}: setter for custom.expr

\item \code{cell.locs<-}: setter for cell.locs

\item \code{cell.locs<-,BanksyObject,assay-method}: setter for cell.locs

\item \code{meta.data<-}: setter for meta.data

\item \code{meta.data<-,BanksyObject,data.frame-method}: setter for meta.data

\item \code{dim.reduction<-}: setter for dim.reduction

\item \code{dim.reduction<-,BanksyObject,list-method}: setter for dim.reduction

\item \code{head}: head for banksy object
}}

\section{Slots}{

\describe{
\item{\code{own.expr}}{own expression}

\item{\code{nbr.expr}}{neighbour expression}

\item{\code{own.norm.scaled.expr}}{own norm/scaled expression}

\item{\code{nbr.norm.scaled.expr}}{neighbour norm/scaled expression}

\item{\code{custom.expr}}{custom expression}

\item{\code{cell.locs}}{cell locations}

\item{\code{meta.data}}{metadata}

\item{\code{dim.reduction}}{dimension reductions}
}}

