% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BanksyObject.R
\docType{class}
\name{BanksyObject-class}
\alias{BanksyObject-class}
\alias{BanksyObject}
\alias{own.expr}
\alias{own.expr,BanksyObject-method}
\alias{nbr.expr}
\alias{nbr.expr,BanksyObject-method}
\alias{custom.expr}
\alias{custom.expr,BanksyObject-method}
\alias{cell.locs}
\alias{cell.locs,BanksyObject-method}
\alias{meta.data}
\alias{meta.data,BanksyObject-method}
\alias{reduction}
\alias{reduction,BanksyObject-method}
\alias{clust.names}
\alias{clust.names,BanksyObject-method}
\alias{own.expr<-}
\alias{own.expr<-,BanksyObject,assay-method}
\alias{nbr.expr<-}
\alias{nbr.expr<-,BanksyObject,assay-method}
\alias{custom.expr<-}
\alias{custom.expr<-,BanksyObject,assay-method}
\alias{cell.locs<-}
\alias{cell.locs<-,BanksyObject,assay-method}
\alias{meta.data<-}
\alias{meta.data<-,BanksyObject,data.frame-method}
\alias{reduction<-}
\alias{reduction<-,BanksyObject,list-method}
\title{S4 class BanksyObject}
\usage{
BanksyObject(
  own.expr = NULL,
  nbr.expr = NULL,
  custom.expr = NULL,
  cell.locs = NULL,
  meta.data = NULL,
  reduction = NULL,
  genes.filter = "intersect",
  min.cells.expressed = -1
)

own.expr(object)

\S4method{own.expr}{BanksyObject}(object)

nbr.expr(object)

\S4method{nbr.expr}{BanksyObject}(object)

custom.expr(object)

\S4method{custom.expr}{BanksyObject}(object)

cell.locs(object)

\S4method{cell.locs}{BanksyObject}(object)

meta.data(object)

\S4method{meta.data}{BanksyObject}(object)

reduction(object)

\S4method{reduction}{BanksyObject}(object)

clust.names(object)

\S4method{clust.names}{BanksyObject}(object)

own.expr(object) <- value

\S4method{own.expr}{BanksyObject,assay}(object) <- value

nbr.expr(object) <- value

\S4method{nbr.expr}{BanksyObject,assay}(object) <- value

custom.expr(object) <- value

\S4method{custom.expr}{BanksyObject,assay}(object) <- value

cell.locs(object) <- value

\S4method{cell.locs}{BanksyObject,assay}(object) <- value

meta.data(object) <- value

\S4method{meta.data}{BanksyObject,data.frame}(object) <- value

reduction(object) <- value

\S4method{reduction}{BanksyObject,list}(object) <- value
}
\arguments{
\item{own.expr}{own expression}

\item{nbr.expr}{neighbour expression}

\item{custom.expr}{custom expression}

\item{cell.locs}{cell locations}

\item{meta.data}{metadata}

\item{reduction}{dimension reductions}

\item{genes.filter}{intersect or union}

\item{min.cells.expressed}{min number of cells a gene is expressed in
(to be used with intersect)}

\item{object}{BanksyObject}

\item{value}{to be assigned}
}
\value{
BanksyObject
}
\description{
S4 class BanksyObject
}
\section{Functions}{
\itemize{
\item \code{own.expr}: getter own.expr

\item \code{own.expr,BanksyObject-method}: getter own.expr

\item \code{nbr.expr}: getter nbr.expr

\item \code{nbr.expr,BanksyObject-method}: getter nbr.expr

\item \code{custom.expr}: getter custom.expr

\item \code{custom.expr,BanksyObject-method}: getter custom.expr

\item \code{cell.locs}: getter cell.locs

\item \code{cell.locs,BanksyObject-method}: getter cell.locs

\item \code{meta.data}: getter meta.data

\item \code{meta.data,BanksyObject-method}: getter meta.data

\item \code{reduction}: getter reduction

\item \code{reduction,BanksyObject-method}: getter reduction

\item \code{clust.names}: getter clust.names

\item \code{clust.names,BanksyObject-method}: getter clust.names

\item \code{own.expr<-}: setter for own.expr

\item \code{own.expr<-,BanksyObject,assay-method}: setter for own.expr

\item \code{nbr.expr<-}: setter for nbr.expr

\item \code{nbr.expr<-,BanksyObject,assay-method}: setter for nbr.expr

\item \code{custom.expr<-}: setter for custom.expr

\item \code{custom.expr<-,BanksyObject,assay-method}: setter for custom.expr

\item \code{cell.locs<-}: setter for cell.locs

\item \code{cell.locs<-,BanksyObject,assay-method}: setter for cell.locs

\item \code{meta.data<-}: setter for meta.data

\item \code{meta.data<-,BanksyObject,data.frame-method}: setter for meta.data

\item \code{reduction<-}: setter for reduction

\item \code{reduction<-,BanksyObject,list-method}: setter for reduction
}}

\section{Slots}{

\describe{
\item{\code{own.expr}}{own expression}

\item{\code{nbr.expr}}{neighbour expression}

\item{\code{custom.expr}}{custom expression}

\item{\code{cell.locs}}{cell locations}

\item{\code{meta.data}}{metadata}

\item{\code{reduction}}{dimension reductions}
}}

