% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ClusterBanksy}
\alias{ClusterBanksy}
\title{Cluster a BANKSY embedding}
\usage{
ClusterBanksy(
  bank,
  lambda = 0.2,
  use_agf = TRUE,
  pca = TRUE,
  npcs = 20,
  method = c("leiden", "louvain", "mclust", "kmeans"),
  k.neighbors = 50,
  resolution = 1,
  leiden.iter = -1,
  num.cores = 1,
  mclust.G = NULL,
  kmeans.centers = NULL,
  kmeans.iter.max = 10,
  seed = 1000,
  M = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{(numeric) weighting parameter (default: 0.2 for celltyping, 0.8 for domain finding)}

\item{use_agf}{(logical) vector specifying whether to use the AGF for clustering}

\item{pca}{(logical) runs clustering on PCs (TRUE) or BANKSY matrix (FALSE)}

\item{npcs}{(integer) number of pcs to use for clustering (default: 20)}

\item{method}{(character) one of leiden, louvain, mclust, kmeans}

\item{k.neighbors}{(integer) parameter for constructing sNN (for louvain / leiden)}

\item{resolution}{(numeric) parameter used for clustering (leiden)}

\item{leiden.iter}{(integer) number of leiden iterations (leiden)}

\item{num.cores}{(integer) number of parallel cores on unix / macOS platforms (leiden)}

\item{mclust.G}{(integer) number of mixture components (mclust)}

\item{kmeans.centers}{(integer) number of clusters (kmeans)}

\item{kmeans.iter.max}{(integer) max number of iterations (kmeans)}

\item{seed}{(integer) random seed for reproducibility}

\item{M}{(integer) advanced usage. vector specifying the highest azimuthal
Fourier harmonic to cluster with. If specified, overwrites the \code{use_agf} argument.}

\item{verbose}{(logical) show progress bar}

\item{...}{to pass to methods}
}
\value{
BanksyObject with cluster labels in meta.data
}
\description{
Cluster the BANKSY embedding obtained from different combinations of 
parameters (`lambda`, `use_agf`, `npcs`). Multiple clustering methods are 
implemented, with the default being Leiden graph-based clustering. For this 
method, parallelization is implemented on non-Windows operating systems.

This function allows a grid search over multiple parameters. Parameters 
which admit vectors are:
\itemize{
  \item{`lambda`}
  \item{`use_agf`}
  \item{`k.neighbors`}
  \item{`mclust.G`}
  \item{`kmeans.centers`}
}
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- RunBanksyPCA(bank, lambda = 0.3)
set.seed(100)
bank <- ClusterBanksy(bank, lambda = 0.3, npcs = 20, k.neighbors = 50, resolution = 0.5)

}
