% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ClusterBanksy}
\alias{ClusterBanksy}
\title{Cluster based on joint expression matrix}
\usage{
ClusterBanksy(
  bank,
  lambda = 0.25,
  resolution = 0.8,
  kneighbours = 40,
  npcs = 50,
  nneighbors = 30,
  spread = 3,
  nepochs = 300,
  mindist = 0.3,
  leideniters = -1,
  verbose = FALSE
)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{weighting parameter - larger values incorporate more spatial
information}

\item{resolution}{parameter used for clustering}

\item{kneighbours}{parameter for constructing shared nearest neighbor network}

\item{npcs}{number of principal components to use}

\item{nneighbors}{umap parameter - number of neighbors to use for umap}

\item{spread}{umap parameter - effective scale of embedded points}

\item{nepochs}{umap parameter - number of epochs to run umap optimization}

\item{mindist}{umap parameter - effective min. dist. between embedded points}

\item{leideniters}{number of leiden iterations}

\item{verbose}{output messages}
}
\value{
BanksyObject with cluster labels in meta.data
}
\description{
Cluster based on joint expression matrix
}
