% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ClusterBanksy}
\alias{ClusterBanksy}
\title{Cluster based on joint expression matrix}
\usage{
ClusterBanksy(
  bank,
  lambda,
  K = 1,
  pca = TRUE,
  npcs = 30,
  method = c("leiden", "louvain", "mclust", "kmeans"),
  k.neighbors = 50,
  resolution = 1,
  leiden.iter = -1,
  mclust.G = NULL,
  kmeans.centers = NULL,
  kmeans.iter.max = 10,
  seed = 42,
  ...
)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{(numeric) weighting parameter}

\item{K}{(numeric) compute up to the k-th azimuthal fourier harmonic (default: 1)}

\item{pca}{(logical) runs clustering on PCs (TRUE) or BANKSY matrix (FALSE)}

\item{npcs}{(numeric) number of pcs to use for clustering}

\item{method}{(character) one of leiden, louvain, mclust, kmeans}

\item{k.neighbors}{(numeric) parameter for constructing sNN (for louvain / leiden)}

\item{resolution}{(numeric) parameter used for clustering (leiden)}

\item{leiden.iter}{(numeric) number of leiden iterations (leiden)}

\item{mclust.G}{(numeric) number of mixture components (mclust)}

\item{kmeans.centers}{(numeric) number of clusters (kmeans)}

\item{kmeans.iter.max}{(numeric) max number of iterations (kmeans)}

\item{seed}{seed}

\item{...}{to pass to methods}
}
\value{
BanksyObject with cluster labels in meta.data
}
\description{
Cluster based on joint expression matrix
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- RunPCA(bank, lambda = 0.3)
bank <- ClusterBanksy(bank, lambda = 0.3, npcs = 20, k.neighbors = 50, resolution = 0.5)

}
