% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterBanksy}
\alias{clusterBanksy}
\title{Cluster based on joint expression matrix}
\usage{
clusterBanksy(
  se,
  use_agf = TRUE,
  lambda = 0.2,
  algo = "leiden",
  use_pcs = TRUE,
  npcs = 20,
  assay_name = NULL,
  group = NULL,
  k_neighbors = 50,
  resolution = 1,
  leiden.iter = -1,
  M = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{se}{A \code{SpatialExperiment},
\code{SingleCellExperiment} or \code{SummarizedExperiment}
  object with \code{computeBanksy} ran.}

\item{use_agf}{A logical vector specifying whether to use the AGF for 
clustering.}

\item{lambda}{A numeric scalar in \eqn{\in [0,1]} specifying a spatial
weighting parameter. Larger values incorporate more spatial neighborhood
information.}

\item{algo}{A string scalar specifying the clustering algorithm to use; one
of leiden, louvain, mclust, kmeans.}

\item{use_pcs}{A logical scalar specifying whether to cluster on PCs. If
FALSE, runs on the BANKSY matrix.}

\item{npcs}{An integer scalar specifying the number of principal components
to use if \code{use_pcs} is TRUE.}

\item{assay_name}{A string scalar specifying the name of the assay used in
\code{computeBanksy}.}

\item{group}{A string scalar specifying a grouping variable for samples in
\code{se}. This is used to scale the samples in each group separately.}

\item{k_neighbors}{An integer scalar specifying number of neighbors for
constructing sNN (for louvain / leiden).}

\item{resolution}{A numeric scalar specifying resolution used for clustering
(leiden).}

\item{leiden.iter}{An integer scalar specifying the number of leiden
iterations. For running till convergence, set to -1 (leiden).}

\item{M}{Advanced usage. An integer vector specifying the highest azimuthal
Fourier harmonic to cluster with. If specified, overwrites the 
\code{use_agf} argument.}

\item{seed}{Random seed.}

\item{...}{to pass to methods}
}
\value{
A SpatialExperiment / SingleCellExperiment / SummarizedExperiment
  object with cluster labels in \code{colData(se)}.
}
\description{
Cluster based on joint expression matrix
}
\examples{
data(rings)
spe <- computeBanksy(rings, assay_name = "counts", M = 1, k_geom = c(15, 30))
spe <- runBanksyPCA(spe, M = 1, lambda = c(0, 0.2), npcs = 20)
spe <- clusterBanksy(spe, M = 1, lambda = c(0, 0.2), resolution = 1)

}
