% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ClusterBanksy}
\alias{ClusterBanksy}
\title{Cluster based on joint expression matrix}
\usage{
ClusterBanksy(
  bank,
  lambda = 0.5,
  resolution = 1.2,
  kneighbours = 30,
  npcs = 50,
  leiden_iters = -1,
  pypath = "/usr/bin/python3",
  use_docker = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{lambda mixing}

\item{resolution}{resolution}

\item{kneighbours}{kneighbours (sNN)}

\item{npcs}{num princ. comp.}

\item{leiden_iters}{leiden iters}

\item{pypath}{pypath}

\item{use_docker}{use docker}

\item{verbose}{messages}
}
\value{
BanksyObject with cluster labels in meta.data
}
\description{
Cluster based on joint expression matrix
}
