% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ClusterBanksy}
\alias{ClusterBanksy}
\title{Cluster based on joint expression matrix}
\usage{
ClusterBanksy(
  bank,
  lambda = 0.25,
  resolution = 0.8,
  kneighbours = 40,
  npcs = 50,
  nneighbors = 30,
  spread = 3,
  nepochs = 300,
  mindist = 0.3,
  leideniters = -1,
  verbose = FALSE
)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{lambda mixing}

\item{resolution}{resolution}

\item{kneighbours}{kneighbours (sNN)}

\item{npcs}{num princ. comp.}

\item{nneighbors}{umap}

\item{spread}{umap}

\item{nepochs}{umap}

\item{mindist}{umap}

\item{leideniters}{leiden iters}

\item{verbose}{messages}
}
\value{
BanksyObject with cluster labels in meta.data
}
\description{
Cluster based on joint expression matrix
}
