% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ComputeBanksy}
\alias{ComputeBanksy}
\title{Compute Banksy Matrices}
\usage{
ComputeBanksy(
  bank,
  sigma = 1.5,
  alpha = 0.05,
  kspatial = 1000,
  dimensions = "all",
  spatialMode = "kNN_r",
  k_geom = 10
)
}
\arguments{
\item{bank}{BanksyObject}

\item{sigma}{sigma}

\item{alpha}{alpha}

\item{kspatial}{kspatial}

\item{dimensions}{dims}

\item{spatialMode}{spatialMode}

\item{k_geom}{k_geom}
}
\value{
BanksyObject
}
\description{
Compute Banksy Matrices
}
