% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ComputeBanksy}
\alias{ComputeBanksy}
\title{Compute Banksy Matrices}
\usage{
ComputeBanksy(
  bank,
  normalizeColumns = TRUE,
  normalizeColumnsTo = 100,
  zScaleRows = TRUE,
  zScaleBeforeAveraging = FALSE,
  zScaleOwnAfterAveraging = TRUE,
  zScaleNbrAfterAveraging = TRUE,
  sigma = 1.5,
  alpha = 0.05,
  kspatial = 1000,
  dimensions = "all",
  spatialMode = "kNN_r",
  k_geom = 10
)
}
\arguments{
\item{bank}{BanksyObject}

\item{normalizeColumns}{to normalize columns}

\item{normalizeColumnsTo}{a scale factor}

\item{zScaleRows}{to zscale rows}

\item{zScaleBeforeAveraging}{to zscale before avg}

\item{zScaleOwnAfterAveraging}{to zscale own after avg}

\item{zScaleNbrAfterAveraging}{to zscale nbr after avg}

\item{sigma}{sigma}

\item{alpha}{alpha}

\item{kspatial}{kspatial}

\item{dimensions}{dims}

\item{spatialMode}{spatialMode}

\item{k_geom}{k_geom}
}
\value{
Banksy Object
}
\description{
Compute Banksy Matrices
}
