% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{ComputeBanksy}
\alias{ComputeBanksy}
\title{Compute Banksy Matrices}
\usage{
ComputeBanksy(
  bank,
  compute_agf = FALSE,
  spatial_mode = "kNN_median",
  k_geom = 15,
  n = 2,
  sigma = 1.5,
  alpha = 0.05,
  k_spatial = 100,
  M = NULL,
  sample_size = NULL,
  sample_renorm = FALSE,
  seed = NULL,
  dimensions = "all",
  center = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{bank}{BanksyObject}

\item{compute_agf}{(logical) whether to compute the AGF}

\item{spatial_mode}{(character) 
\itemize{
 \item{kNN_r: k-nearest neighbors with $1/r$ kernel}
 \item{kNN_rn: k-nearest neighbors with $1/r^n$ kernel}
 \item{kNN_rank: k-nearest neighbors with rank Gaussian kernel}
 \item{kNN_unif: k-nearest neighbors wth uniform kernel}
 \item{kNN_median: k-nearest neighbors with median-scaled Gaussian kernel (default)}
 \item{rNN_gauss: radial nearest neighbors with Gaussian kernel}
}}

\item{k_geom}{(numeric) number of neighbors to use (for kNN)}

\item{n}{(numeric) exponent of radius (for kNN_rn)}

\item{sigma}{(numeric) std. dev. of Gaussian kernel (for rNN_gauss)}

\item{alpha}{(numeric) determines radius used: larger alphas give
smaller radii (for rNN_gauss)}

\item{k_spatial}{(numeric) initial number of neighbors to use (for rNN_gauss)}

\item{M}{(numeric) advanced usage. specifies the highest azimuthal Fourier
harmonic to compute. if specified, overwrites the \code{use_agf} argument}

\item{sample_size}{(numeric) number of neighbors to sample from the neighborhood}

\item{sample_renorm}{(logical) whether to renormalize the neighbor weights to 1}

\item{seed}{(numeric) seed for sampling the neighborhood}

\item{dimensions}{(character vector) dimensions to use when computing neighborhood
\itemize{
 \item{subset of colnames of cell.locs}
 \item{all}{Uses all colnames of cell.locs to compute (default)}
}}

\item{center}{(logical) center higher order harmonics in local neighborhoods}

\item{verbose}{messages}
}
\value{
BanksyObject
}
\description{
Compute Banksy Matrices
}
\examples{
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
# Compute neighbors 
bank <- ComputeBanksy(bank)

}
