% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ConnectClusters}
\alias{ConnectClusters}
\title{Harmonize cluster labels among parameter runs}
\usage{
ConnectClusters(bank, map.to = NULL)
}
\arguments{
\item{bank}{Banksy Object}

\item{map.to}{(character) specify a cluster to map to}
}
\value{
BanksyObject with harmonized cluster labels
}
\description{
Harmonize cluster labels among parameter runs
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunPCA(bank, lambda = 0.2)
bank <- ClusterBanksy(bank, lambda = 0.2, npcs = 20, k.neighbors = 50, resolution = c(0.5,1.5))
bank <- ConnectClusters(bank)

}
