% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ConnectClusters}
\alias{ConnectClusters}
\title{Harmonise cluster labels among parameter runs}
\usage{
ConnectClusters(bank, mapto = NULL, reverse = FALSE)
}
\arguments{
\item{bank}{Banksy Object}

\item{mapto}{specify a cluster to map to}

\item{reverse}{reverse the map}
}
\value{
A list with BanksyObject with harmonised cluster labels and
  adjusted Rand indices for each cluster mapping
}
\description{
Harmonise cluster labels among parameter runs
}
