% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{ConnectClusters}
\alias{ConnectClusters}
\title{Harmonize cluster labels among parameter runs}
\usage{
ConnectClusters(bank, map.to = NULL)
}
\arguments{
\item{bank}{Banksy Object}

\item{map.to}{specify a cluster to map to}
}
\value{
BanksyObject with harmonized cluster labels
}
\description{
Harmonize cluster labels among parameter runs
}
