% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ConnectClusters}
\alias{ConnectClusters}
\title{Harmonise cluster labels among parameter runs}
\usage{
ConnectClusters(bank, verbose = FALSE, optim = TRUE)
}
\arguments{
\item{bank}{Banksy Object}

\item{verbose}{output messages}

\item{optim}{optimize cluster mapping based on parent distribution in
ambiguous cases}
}
\value{
BanksyObject with harmonised cluster labels
}
\description{
Harmonise cluster labels among parameter runs
}
