% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{ConnectClusters}
\alias{ConnectClusters}
\title{Harmonise cluster labels among parameter runs}
\usage{
ConnectClusters(bank, verbose = FALSE, optim = TRUE)
}
\arguments{
\item{bank}{Banksy Object}

\item{verbose}{verbose or not}

\item{optim}{optimise or not (KS-test)}
}
\value{
BanksyObject with harmonised cluster labels
}
\description{
Harmonise cluster labels among parameter runs
}
