% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getARI}
\alias{getARI}
\title{Calculate adjusted rand index for harmonized clusters}
\usage{
getARI(bank, digits = 3)
}
\arguments{
\item{bank}{BanksyObject with harmonized clusters}

\item{digits}{(numeric) number of digits to round ARI to}
}
\value{
matrix of ARI
}
\description{
Calculate adjusted rand index for harmonized clusters
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunBanksyPCA(bank, lambda = 0.2)
bank <- ClusterBanksy(bank, lambda = 0.2, npcs = 20, k.neighbors = 50, resolution = c(0.5,1.5))
ari <- getARI(bank)
ari

}
