% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{getBanksyMatrix}
\alias{getBanksyMatrix}
\title{Returns the BANKSY matrix}
\usage{
getBanksyMatrix(
  se,
  M,
  lambda,
  assay_name = NULL,
  scale = FALSE,
  group = NULL,
  verbose = TRUE
)
}
\arguments{
\item{se}{A \code{SpatialExperiment},
\code{SingleCellExperiment} or \code{SummarizedExperiment}
  object with \code{computeBanksy} ran.}

\item{M}{A integer scalar specifying the highest azimuthal
Fourier harmonic to compute.}

\item{lambda}{A numeric scalar \eqn{\in [0,1]} specifying a spatial
weighting parameter. Larger values incorporate more spatial neighborhood
information.}

\item{assay_name}{A string scalar specifying the name of the assay used in
\code{computeBanksy}.}

\item{scale}{A logical scalar specifying whether to scale the features to
zero mean and unit standard deviation. This is performed before
multiplying the assays by their corresponding lambda weighting factors.}

\item{group}{A string scalar specifying a grouping variable for samples in
\code{se}. This is used to scale the samples in each group separately.}

\item{verbose}{A logical scalar specifying verbosity.}
}
\value{
BANKSY matrix.
}
\description{
Returns the BANKSY matrix
}
\examples{
data(rings)
spe <- computeBanksy(rings, assay_name = "counts", M = 1, k_geom = c(15, 30))
banksyMatrix <- getBanksyMatrix(spe, M = 1, lambda = 0.2)

}
