% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getBanksyMatrix}
\alias{getBanksyMatrix}
\title{Returns the Banksy matrix (own + nbr)}
\usage{
getBanksyMatrix(bank, lambda = 0.25, transformed = TRUE)
}
\arguments{
\item{bank}{Banksy Object}

\item{lambda}{lambda}

\item{transformed}{return the normalized scaled matrix}
}
\value{
BanksyMatrix
}
\description{
Returns the Banksy matrix (own + nbr)
}
