% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{getBanksyMatrix}
\alias{getBanksyMatrix}
\title{Returns the Banksy matrix (own + nbr)}
\usage{
getBanksyMatrix(bank, lambda = 0.25)
}
\arguments{
\item{bank}{Banksy Object}

\item{lambda}{weighting parameter - larger values incorporate more spatial
information}
}
\value{
BanksyMatrix
}
\description{
Returns the Banksy matrix (own + nbr)
}
