% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{getBanksyMatrix}
\alias{getBanksyMatrix}
\title{Returns the Banksy matrix (own + nbr)}
\usage{
getBanksyMatrix(bank, lambda = 0.2, M = 0, verbose = TRUE)
}
\arguments{
\item{bank}{Banksy Object}

\item{lambda}{(numeric) spatial weighting parameter}

\item{M}{(numeric) compute up to the k-th azimuthal fourier harmonic (default: 1)}

\item{verbose}{(logical) verbosity}
}
\value{
BanksyMatrix
}
\description{
Returns the Banksy matrix (own + nbr)
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
# Normalize the own.expr matrix
bank <- NormalizeBanksy(bank)
# Compute BANKSY matrix
bank <- ComputeBanksy(bank)
bm <- getBanksyMatrix(bank)

}
