% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hippocampus}
\alias{hippocampus}
\title{Mouse Hippocampus VeraFISH data}
\format{
A list with 2 entries:
\describe{
\item{expression}{(matrix) gene-cell matrix}
\item{locations}{data.frame} cell centroids in 2D)
}
}
\usage{
hippocampus
}
\description{
This dataset comprises VeraFISH profiling of cells in the mouse hippocampus.
Gene expression and cell centroids for 7,944 cells and 124 genes in 2 spatial
dimensions are provided.
}
\keyword{datasets}
