% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NormalizeBanksy}
\alias{NormalizeBanksy}
\title{Normalize columns to a normFactor}
\usage{
NormalizeBanksy(bank, assay = "both", normFactor = 100)
}
\arguments{
\item{bank}{BanksyObject}

\item{assay}{assay to scale - one of both, own, nbr (default: both)}

\item{normFactor}{normalization factor}
}
\value{
normalized Banksy object
}
\description{
Normalize columns to a normFactor
}
