% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{NormalizeBanksy}
\alias{NormalizeBanksy}
\title{Normalize columns to a normFactor}
\usage{
NormalizeBanksy(
  bank,
  assay = "both",
  norm_factor = 100,
  log_norm = FALSE,
  pseudocount = 0.1,
  base = 10
)
}
\arguments{
\item{bank}{BanksyObject}

\item{assay}{(character) assay to scale
\itemize{
 \item{own: Scales own.expr}
 \item{nbr: Scales nbr.expr}
 \item{both: Scales own.expr and nbr.expr (default)}
}}

\item{norm_factor}{(numeric) normalization factor (default: 100)}

\item{log_norm}{(logical) log transforms data (default: FALSE)}

\item{pseudocount}{(numeric) pseudocount for log transform (default: 0.1)}

\item{base}{(numeric) base for log transform (default: 10)}
}
\value{
normalized Banksy object
}
\description{
Normalize columns to a normFactor
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
# Normalize the own.expr matrix
bank <- NormalizeBanksy(bank)

}
