% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAlluvia}
\alias{plotAlluvia}
\title{Plot Alluvial plot for different cluster assignments}
\usage{
plotAlluvia(bank, max.cells = 500, flow.colors = NULL, cluster.order = NULL)
}
\arguments{
\item{bank}{BanksyObject}

\item{max.cells}{Number of cells to display}

\item{flow.colors}{Colors for alluvial flow}

\item{cluster.order}{Order for clustering runs in horizontal axis}
}
\value{
Alluvial plot
}
\description{
Plot Alluvial plot for different cluster assignments
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunBanksyPCA(bank, lambda = 0.2)
bank <- ClusterBanksy(bank, lambda = 0.2, npcs = 20, k.neighbors = 50, resolution = c(0.5,1.5))
set.seed(100)
plotAlluvia(bank, max.cells = 20)
plotAlluvia(bank, max.cells = 20, cluster.order = rev(clust.names(bank)))

}
