% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap (wrapper for ComplexHeatmap)}
\usage{
plotHeatmap(
  bank,
  assay = "own.expr",
  dataset = NULL,
  lambda = NULL,
  col = NULL,
  col.breaks = NULL,
  cluster.row = TRUE,
  cluster.column = FALSE,
  row.dend = FALSE,
  column.dend = FALSE,
  cex.row = 4,
  cex.column = 4,
  annotate = FALSE,
  annotate.by = NULL,
  annotate.col = NULL,
  name = "Expression",
  ...
)
}
\arguments{
\item{bank}{BanksyObject}

\item{assay}{assay to plot heatmap}

\item{dataset}{dataset to plot heatmap}

\item{lambda}{lambda if assay == banksy}

\item{col}{colours to use in heatmap}

\item{col.breaks}{color breaks to use in heatmap (same number as col is
specified)}

\item{cluster.row}{cluster rows}

\item{cluster.column}{cluster columns}

\item{row.dend}{draw row dendrograms}

\item{column.dend}{draw column dendrograms}

\item{cex.row}{row label size}

\item{cex.column}{column label size}

\item{annotate}{add annotation for cells}

\item{annotate.by}{metadata to annotate cells by}

\item{annotate.col}{colours to annotate cells by}

\item{name}{name of heatmap legend}

\item{...}{parameters to pass to ComplexHeatmap::Heatmap}
}
\value{
Heatmap of class ComplexHeatmap
}
\description{
Plot Heatmap (wrapper for ComplexHeatmap)
}
