% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap (wrapper for ComplexHeatmap)}
\usage{
plotHeatmap(
  bank,
  assay = "own.expr",
  dataset = NULL,
  lambda = NULL,
  cells = NULL,
  features = NULL,
  col = NULL,
  col.breaks = NULL,
  col.discrete = NULL,
  annotate = FALSE,
  annotate.by = NULL,
  order.by = NULL,
  annotation.name = TRUE,
  annotation.size = 6,
  annotation.pos = "right",
  barplot.by = NULL,
  barplot.border = FALSE,
  barplot.width = 0.5,
  cluster_row_slices = FALSE,
  show_row_dend = FALSE,
  show_row_names = TRUE,
  cluster_rows = FALSE,
  cex.row = 4,
  row_gap = 1,
  cluster_column_slices = FALSE,
  show_column_dend = FALSE,
  show_column_names = FALSE,
  cluster_columns = TRUE,
  cex.column = 2,
  column_gap = 1,
  rasterize = FALSE,
  max.cols = NULL,
  seed = 42,
  name = "Expression",
  ...
)
}
\arguments{
\item{bank}{BanksyObject}

\item{assay}{assay to plot heatmap (one of own.expr, nbr.expr or banksy)}

\item{dataset}{dataset to plot heatmap}

\item{lambda}{lambda if assay == banksy}

\item{cells}{specific cells to plot}

\item{features}{specific features to plot}

\item{col}{colours to use in heatmap}

\item{col.breaks}{color breaks to use in heatmap (same number as col is specified)}

\item{col.discrete}{cluster colors, named color array, names are clusters, values are colors in hexadecimal}

\item{annotate}{add annotation for cells}

\item{annotate.by}{metadata to annotate cells by}

\item{order.by}{metadata to order cells by (one of annotate.by)}

\item{annotation.name}{show annotation name}

\item{annotation.size}{size of annotation labels}

\item{annotation.pos}{position of annotation labels}

\item{barplot.by}{metadata to plot barplots (numeric)}

\item{barplot.border}{show borders of barplot}

\item{barplot.width}{barplot width}

\item{cluster_row_slices}{cluster row slices (BANKSY)}

\item{show_row_dend}{show row dendrograms}

\item{show_row_names}{show row names}

\item{cluster_rows}{cluster rows}

\item{cex.row}{row label size}

\item{row_gap}{row gaps between row slices}

\item{cluster_column_slices}{cluster column slices (annotations)}

\item{show_column_dend}{show column dendrograms}

\item{show_column_names}{show column names}

\item{cluster_columns}{cluster columns}

\item{cex.column}{column label size}

\item{column_gap}{column gaps between column slices}

\item{rasterize}{rasterize if TRUE}

\item{max.cols}{max columns to display - will subsample if not NULL}

\item{seed}{used for sampling}

\item{name}{name of heatmap legend}

\item{...}{parameters to pass to ComplexHeatmap::Heatmap}
}
\value{
Heatmap of class ComplexHeatmap
}
\description{
Plot Heatmap (wrapper for ComplexHeatmap)
}
\examples{
# Generate a simulated dataset
d <- simulateDataset(n_cells = c(20,20,20), n_genes = 30)
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
plotHeatmap(bank, annotate = TRUE, annotate.by = 'Label')

}
