% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{plotScree}
\alias{plotScree}
\title{Plots scree plot for PCA}
\usage{
plotScree(bank, lambda)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{(numeric) spatial weighting parameter}
}
\value{
Scree plot
}
\description{
Plots scree plot for PCA
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunPCA(bank, lambda = 0.2)
scree <- plotScree(bank, lambda = 0.2)
scree

}
