% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpatialDims}
\alias{plotSpatialDims}
\title{Plot Spatial dims}
\usage{
plotSpatialDims(
  bank,
  by,
  dataset = NULL,
  col = NULL,
  legend = TRUE,
  main = NULL,
  pt.size = 1,
  main.size = 5,
  legend.text.size = 1,
  legend.pt.size = 1
)
}
\arguments{
\item{bank}{BanksyObject}

\item{by}{clustering run to plot by}

\item{dataset}{if multiple dataset are run}

\item{col}{colours}

\item{legend}{show legend}

\item{main}{title}

\item{pt.size}{size of points}

\item{main.size}{size of title}

\item{legend.text.size}{size of legend text}

\item{legend.pt.size}{size of legend point}
}
\description{
Plot Spatial dims
}
