% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpatialDims}
\alias{plotSpatialDims}
\title{Plot Spatial dims}
\usage{
plotSpatialDims(bank, params, col = NULL, main = NULL, pt.size = 1, main.size)
}
\arguments{
\item{bank}{BanksyObject}

\item{params}{clustering run to plot by}

\item{col}{colours}

\item{main}{title}

\item{pt.size}{size of points}

\item{main.size}{size of title}
}
\description{
Plot Spatial dims
}
