% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSpatialFeatures}
\alias{plotSpatialFeatures}
\title{Wrapper for plotSpatial}
\usage{
plotSpatialFeatures(bank, dataset = NULL, by, type, nrow, ncol, ...)
}
\arguments{
\item{bank}{BanksyObject}

\item{dataset}{if multiple datasets are used}

\item{by}{features to plot}

\item{type}{type corresponding to features}

\item{nrow}{nrow for grob}

\item{ncol}{ncol for grob}

\item{...}{parameters to pass to plotSpatial}
}
\description{
Wrapper for plotSpatial
}
