% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP}
\usage{
plotUMAP(bank, params, col = NULL, legend = TRUE, main = NULL, size = 1)
}
\arguments{
\item{bank}{BanksyObject}

\item{params}{clustering run to plot by}

\item{col}{cluster colours}

\item{legend}{cluster legend}

\item{main}{title}

\item{size}{size of points}
}
\description{
Plot UMAP
}
