% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP}
\usage{
plotUMAP(
  bank,
  by,
  reduction,
  col = NULL,
  legend = TRUE,
  main = NULL,
  pt.size = 1,
  main.size = 5,
  legend.text.size = 6,
  legend.pt.size = 3
)
}
\arguments{
\item{bank}{BanksyObject}

\item{by}{groupings for cells}

\item{reduction}{reduction to use}

\item{col}{cluster colours}

\item{legend}{cluster legend}

\item{main}{title}

\item{pt.size}{size of points}

\item{main.size}{size of title}

\item{legend.text.size}{size of legend text}

\item{legend.pt.size}{size of legend point}
}
\description{
Plot UMAP
}
