% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rings}
\alias{rings}
\title{An unrealistic simulation of spatially-resolved omics data.}
\format{
A SpatialExperiment object.
}
\usage{
data(rings)
}
\value{
A SpatialExperiment object
}
\description{
This dataset comprises gene expression and spatial coordinates for 50 genes
and 308 cells from 4 clusters (\code{rings$clusters}). See \code{system.file('scripts/rings.R', package='Banksy')} on how this dataset was generated.
}
\keyword{datasets}
