% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{RunBanksyPCA}
\alias{RunBanksyPCA}
\title{Run PCA}
\usage{
RunBanksyPCA(bank, lambda = 0.2, M = 1, npcs = 20, verbose = TRUE)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{(numeric vector) spatial weighting parameter}

\item{M}{(numeric vector) run PCA using up to the m-th azimuthal fourier harmonic (default: 1)}

\item{npcs}{(numeric) number of principal components to compute}

\item{verbose}{(logical) print messages}
}
\value{
BanksyObject with PCA
}
\description{
Run PCA
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunBanksyPCA(bank, lambda = 0.2)

}
