% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{RunBanksyUMAP}
\alias{RunBanksyUMAP}
\title{Run UMAP}
\usage{
RunBanksyUMAP(
  bank,
  lambda = 0.2,
  M = 1,
  ncomponents = 2,
  pca = TRUE,
  npcs = 20,
  nneighbors = 30,
  spread = 3,
  mindist = 0.1,
  nepochs = 300,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{(numeric vector) spatial weighting parameter}

\item{M}{(numeric) compute up to the k-th azimuthal fourier harmonic (default: 1)}

\item{ncomponents}{(numeric) number of umap components to compute}

\item{pca}{(logical) run UMAP on PCs (TRUE) or BANKSY matrix (FALSE)}

\item{npcs}{(numeric) number of principal components to use for umap}

\item{nneighbors}{(numeric) number of neighbors to use for umap}

\item{spread}{(numeric) effective scale of embedded points}

\item{mindist}{(numeric) effective min. dist. between embedded points}

\item{nepochs}{(numeric) number of epochs to run umap optimization}

\item{verbose}{(logical) print messages}

\item{...}{parameters to pass to uwot::umap}
}
\value{
BanksyObject with dimensionality reduction
}
\description{
Run UMAP
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunBanksyPCA(bank, lambda = 0.2)
bank <- RunBanksyUMAP(bank, lambda = 0.2)

}
