% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{RunUMAP}
\alias{RunUMAP}
\title{Run UMAP}
\usage{
RunUMAP(
  bank,
  lambda,
  ncomponents = 2,
  pca = TRUE,
  npcs = 20,
  nneighbors = 30,
  spread = 3,
  mindist = 0.3,
  nepochs = 300,
  ...
)
}
\arguments{
\item{bank}{BanksyObject}

\item{lambda}{numeric vector - weighting parameter}

\item{ncomponents}{number of umap components to compute}

\item{pca}{if TRUE, run UMAP on pca, if not runs on Banksy matrix}

\item{npcs}{number of principal components to use for umap}

\item{nneighbors}{number of neighbors to use for umap}

\item{spread}{effective scale of embedded points}

\item{mindist}{effective min. dist. between embedded points}

\item{nepochs}{number of epochs to run umap optimization}

\item{...}{parameters to pass to uwot::umap}
}
\value{
BanksyObject with dimensionality reduction
}
\description{
Run UMAP
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
bank <- NormalizeBanksy(bank)
bank <- ScaleBanksy(bank)
bank <- ComputeBanksy(bank)
bank <- RunPCA(bank, lambda = 0.2)
bank <- RunUMAP(bank, lambda = 0.2)

}
