% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ScaleBanksy}
\alias{ScaleBanksy}
\title{Scale rows}
\usage{
ScaleBanksy(bank, separate = TRUE)
}
\arguments{
\item{bank}{BanksyObject}

\item{separate}{scale datasets separately}
}
\value{
scaled BanksyObject
}
\description{
Scale rows
}
