% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computation.R
\name{ScaleBanksy}
\alias{ScaleBanksy}
\title{Scale rows}
\usage{
ScaleBanksy(bank, assay = "both", separate = TRUE)
}
\arguments{
\item{bank}{BanksyObject}

\item{assay}{assay to scale - one of both, own, nbr (default: both)}

\item{separate}{scale datasets separately}
}
\value{
scaled BanksyObject
}
\description{
Scale rows
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
# Normalize the own.expr matrix
bank <- NormalizeBanksy(bank)
# Scale the own.expr matrix
bank <- ScaleBanksy(bank)

}
