% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateDataset}
\alias{simulateDataset}
\title{Simulate a dataset consisting of expression and location information 
with concentric circular structure}
\usage{
simulateDataset(
  n_buds = 3,
  n_genes = 100,
  n_deg = c(10, 5, 7),
  de_shift = c(3, 4, 2),
  de_dropout = c(0.8, 0.9, 0.7),
  n_cells = c(250, 150, 100),
  n_circles = c(3, 3, 3),
  n_thickness = list(c(3, 1, 3), c(2, 1, 4), c(2, 2, 2)),
  radius_init = c(5, 3, 3),
  radius_increment = c(4, 3, 3),
  x_stretch = c(1.1, 1, 1),
  y_stretch = c(0.9, 1.1, 1),
  offset_x = c(0, 50, 20),
  offset_y = c(0, 40, 35),
  skew = c(0, 0, 0.5),
  seed = 1000
)
}
\arguments{
\item{n_buds}{Number of rings}

\item{n_genes}{Number of genes}

\item{n_deg}{Number of de genes per bud}

\item{de_shift}{Mean shift for de genes per bud}

\item{de_dropout}{Fraction of cells where gene is de per bud}

\item{n_cells}{Number of cells per bud}

\item{n_circles}{Number of circles per bud}

\item{n_thickness}{Thickness of each circle per bud}

\item{radius_init}{Initial radius of each bud}

\item{radius_increment}{Radius increment per bud}

\item{x_stretch}{Amount of stretching in the x-direction per bud}

\item{y_stretch}{Amount of stretching in the y-direction per bud}

\item{offset_x}{Offset to shift x center of bud}

\item{offset_y}{Offset to shift y center of bud}

\item{skew}{Rotation of bud}

\item{seed}{Seed for sampling}
}
\description{
Simulate a dataset consisting of expression and location information 
with concentric circular structure
}
