% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{SmoothLabels}
\alias{SmoothLabels}
\title{Label smoothing as described in SpiceMix (https://doi.org/10.1038/s41588-022-01256-z).
Implemented for numeric labels only}
\usage{
SmoothLabels(
  bank,
  cluster_names = NULL,
  k = 15L,
  prop_thres = 0.5,
  max_iter = 10,
  verbose = TRUE,
  include_self = TRUE
)
}
\arguments{
\item{bank}{BanksyObject}

\item{cluster_names}{(character) vector of label names to smooth. If NULL,
smooths labels in clust.names(bank)}

\item{k}{(integer) number of neighbors (default: 10)}

\item{prop_thres}{(numeric) proportions threshold. If the fraction of 
neighbors with a certain label exceeds this proportion, change the label 
of the current sample (default: 0.5)}

\item{max_iter}{(integer) max number of smoothing iterations. Set to -1 for 
smoothing to convergence. (default: 10)}

\item{verbose}{(logical) verbosity}

\item{include_self}{(logical) include the spots label in the neighborhood. 
(default: True)}
}
\value{
BanksyObject with smoothed cluster labels (appended with '_smooth')
  in meta.data
}
\description{
Label smoothing as described in SpiceMix (https://doi.org/10.1038/s41588-022-01256-z).
Implemented for numeric labels only
}
\examples{
data(dlpfc151673)
bank = BanksyObject(
    own.expr = dlpfc151673$expression, 
    cell.locs = dlpfc151673$locations)
n_spots = ncol(own.expr(bank))
labels = readRDS(
    system.file('extdata/dlpfc_clusters.rds', package = 'Banksy')
    )[seq(n_spots)]
meta.data(bank)$clust = labels
bank = SmoothLabels(bank, k=6)
plotSpatialFeatures(bank, pt.size = 1, by = clust.names(bank), 
    type=rep('discrete',2), ncol = 2, main = clust.names(bank))

}
