% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SplitBanksy}
\alias{SplitBanksy}
\title{Split a BanksyObject by metadata column}
\usage{
SplitBanksy(bank, by, names = NULL)
}
\arguments{
\item{bank}{BanksyObject}

\item{by}{metadata column}

\item{names}{names for new datasets}
}
\value{
BanksyObject
}
\description{
Split a BanksyObject by metadata column
}
