% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling.R
\name{SubsetBanksy}
\alias{SubsetBanksy}
\title{Subset a BanksyObject}
\usage{
SubsetBanksy(
  x,
  cells = NULL,
  features = NULL,
  dims = TRUE,
  metadata = TRUE,
  dataset = NULL
)
}
\arguments{
\item{x}{a BanksyObject}

\item{cells}{cells to filter by}

\item{features}{genes to filter by}

\item{dims}{dimensions to filter by - must correspond to valid columns in
cell.locs}

\item{metadata}{metadata to filter by - must be valid column in meta.data}

\item{dataset}{dataset to subset dimensions by}
}
\value{
subset BanksyObject
}
\description{
Subset a BanksyObject
}
\examples{
# Generate a simulated dataset
d <- simulateDataset()
bank <- BanksyObject(own.expr = d$gcm, cell.locs = d$locs, meta.data = d$meta)
# Subset cells with Label = 1
bank <- SubsetBanksy(bank, metadata = Label == 1)

}
