% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SubsetBanksy}
\alias{SubsetBanksy}
\title{Subset a BanksyObject}
\usage{
SubsetBanksy(
  x,
  cells = TRUE,
  features = NULL,
  dimx = TRUE,
  dimy = TRUE,
  dimz = TRUE,
  metadata = TRUE
)
}
\arguments{
\item{x}{a BanksyObject}

\item{cells}{cells to filter by}

\item{features}{genes to filter by}

\item{dimx}{dimx to filter by - must be valid column}

\item{dimy}{dimy to filter by - must be valid column}

\item{dimz}{dimz to filter by - must be valid column}

\item{metadata}{metadata to filter by - must be valid colulmn}
}
\value{
subset BanksyObject
}
\description{
Subset a BanksyObject
}
