% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SubsetBanksy}
\alias{SubsetBanksy}
\title{Subset a BanksyObject}
\usage{
SubsetBanksy(
  x,
  cells = NULL,
  features = NULL,
  dims = TRUE,
  metadata = TRUE,
  dataset = NULL
)
}
\arguments{
\item{x}{a BanksyObject}

\item{cells}{cells to filter by}

\item{features}{genes to filter by}

\item{dims}{dimensions to filter by - must correspond to valid columns}

\item{metadata}{metadata to filter by - must be valid colulmn}

\item{dataset}{dataset to subset dimensions by}
}
\value{
subset BanksyObject
}
\description{
Subset a BanksyObject
}
